Obviously terrible hack to allow loading of datatree built in 64 bits on a 32 bit system.  Need to rewrite so it always uses 32 bits and then get rid of this stuff.

+++ DataTree.h
@@ -121,7 +121,7 @@ namespace KEngineCore
 		std::vector<int>			mInts;
 		std::vector<float>			mFloats;
 		std::vector<uint8_t>		mBitFields;
-		std::vector<size_t>			mStringIndices;
+		std::vector<int64_t>			mStringIndices;
 		std::vector<StringHash>		mHashes;
 		std::vector<DataTree*>		mBranches;
